backend/app/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arch==7.2.0
astropy==7.0.1
astropy-iers-data==0.2025.3.17.0.34.53
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
contourpy==1.3.1
cvxpy==1.6.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==2.12.0
deprecation==2.1.0
empyrical==0.5.5
executing==2.2.0
fastapi==0.115.11
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qdldl==0.1.7.post5
QuantStats==0.0.64
requests==2.32.3
Riskfolio-Lib==7.0.0
scikit-learn==1.6.1
scipy==1.15.2
scs==3.2.7.post2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.6.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==13.1
XlsxWriter==3.2.2
yarl==1.18.3
yfinance==0.2.54
